+++ src/main/java/lermitage/intellij/extra/icons/cfg/dialogs/ModelDialog.java
@@ -7,6 +7,7 @@ import com.intellij.openapi.fileChooser.FileChooserDescriptor;
 import com.intellij.openapi.ui.DialogWrapper;
 import com.intellij.openapi.ui.Messages;
 import com.intellij.openapi.ui.ValidationInfo;
+import com.intellij.openapi.util.IconLoader;
 import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.ui.CheckBoxList;
 import com.intellij.ui.ListUtil;
@@ -15,6 +16,7 @@ import com.intellij.ui.components.JBLabel;
 import com.intellij.ui.components.JBTextField;
 import com.intellij.util.IconUtil;
 import lermitage.intellij.extra.icons.CustomIconLoader;
+import lermitage.intellij.extra.icons.ExtraIconProvider;
 import lermitage.intellij.extra.icons.IJUtils;
 import lermitage.intellij.extra.icons.IconType;
 import lermitage.intellij.extra.icons.Model;
@@ -22,15 +24,23 @@ import lermitage.intellij.extra.icons.ModelCondition;
 import lermitage.intellij.extra.icons.ModelType;
 import lermitage.intellij.extra.icons.cfg.SettingsForm;
 import lermitage.intellij.extra.icons.cfg.SettingsService;
+import lermitage.intellij.extra.icons.providers.Angular2IconProvider;
 import org.jetbrains.annotations.Nullable;
 
+import javax.swing.Icon;
+import javax.swing.ImageIcon;
 import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JComponent;
 import javax.swing.JLabel;
+import javax.swing.JList;
 import javax.swing.JPanel;
+import javax.swing.ListCellRenderer;
 import javax.swing.SwingUtilities;
 import java.awt.BorderLayout;
+import java.awt.Component;
+import java.awt.event.ItemEvent;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -56,6 +66,7 @@ public class ModelDialog extends DialogWrapper {
     private JButton chooseIconButton;
     private JPanel conditionsPanel;
     private JBLabel idLabel;
+    private JComboBox<Object> chooseIconSelector;
 
     private CustomIconLoader.ImageWrapper customIconImage;
     private JPanel toolbarPanel;
@@ -87,6 +98,7 @@ public class ModelDialog extends DialogWrapper {
             try {
                 customIconImage = loadCustomIcon();
                 if (customIconImage != null) {
+                    chooseIconSelector.setSelectedIndex(0);
                     iconLabel.setIcon(IconUtil.createImageIcon(customIconImage.getImage()));
                 }
             } catch (IllegalArgumentException ex) {
@@ -101,6 +113,33 @@ public class ModelDialog extends DialogWrapper {
         for (ModelType value : ModelType.values()) {
             typeComboBox.addItem(value.getFriendlyName());
         }
+
+        chooseIconSelector.addItem("choose custom or bundled icon");
+        List<Model> bundledModels = new ArrayList<>();
+        bundledModels.addAll(ExtraIconProvider.allModels());
+        bundledModels.addAll(Angular2IconProvider.allModels());
+        bundledModels.stream()
+            .map(Model::getIcon)
+            .sorted()
+            .distinct()
+            .forEach(iconPath -> chooseIconSelector.addItem(new BundledIcon(
+                iconPath, "bundled: " + iconPath.replace("/extra-icons/", ""))));
+        ComboBoxRenderer renderer = new ComboBoxRenderer();
+        // customIconImage
+        chooseIconSelector.setRenderer(renderer);
+        chooseIconSelector.setToolTipText("<html>Choose a custom icon with the button on the right,<br>otherwise select a bundled icon in the list.</html>");
+        chooseIconSelector.addItemListener(event -> {
+            if (event.getStateChange() == ItemEvent.SELECTED) {
+                Object item = event.getItem();
+                if (item instanceof BundledIcon) {
+                    BundledIcon bundledIcon = (BundledIcon) item;
+                    iconLabel.setIcon(IconLoader.getIcon((bundledIcon).getIconPath(), CustomIconLoader.class));
+                    customIconImage = new CustomIconLoader.ImageWrapper(bundledIcon.getIconPath());//fixme
+                } else if (item instanceof String) {
+                    iconLabel.setIcon(new ImageIcon());
+                }
+            }
+        });
     }
 
     /**
@@ -110,10 +149,13 @@ public class ModelDialog extends DialogWrapper {
         String icon = null;
         IconType iconType = null;
         if (customIconImage != null) {
-            icon = CustomIconLoader.toBase64(customIconImage);
             iconType = customIconImage.getIconType();
-        }
-        else if (modelToEdit != null) {
+            if (customIconImage.getIconType() == IconType.PATH) {
+                icon = customIconImage.getImageAsBundledIconRef();
+            } else {
+                icon = CustomIconLoader.toBase64(customIconImage);
+            }
+        } else if (modelToEdit != null) {
             icon = modelToEdit.getIcon();
             iconType = modelToEdit.getIconType();
         }
@@ -148,9 +190,17 @@ public class ModelDialog extends DialogWrapper {
         typeComboBox.updateUI();
         Double additionalUIScale = SettingsService.getIDEInstance().getAdditionalUIScale();
         SwingUtilities.invokeLater(() -> iconLabel.setIcon(CustomIconLoader.getIcon(model, additionalUIScale)));
-        model.getConditions().forEach(modelCondition -> {
-            conditionsCheckboxList.addItem(modelCondition, modelCondition.asReadableString(FIELD_SEPARATOR), modelCondition.isEnabled());
-        });
+        if (model.getIconType() == IconType.PATH) {
+            for (int itemIdx = 0; itemIdx < chooseIconSelector.getItemCount(); itemIdx++) {
+                Object item = chooseIconSelector.getItemAt(itemIdx);
+                if (item instanceof BundledIcon && ((BundledIcon)item).iconPath.equals(model.getIcon())) {
+                    chooseIconSelector.setSelectedIndex(itemIdx);
+                    break;
+                }
+            }
+        }
+        model.getConditions().forEach(modelCondition ->
+            conditionsCheckboxList.addItem(modelCondition, modelCondition.asReadableString(FIELD_SEPARATOR), modelCondition.isEnabled()));
     }
 
     /**
@@ -175,9 +225,9 @@ public class ModelDialog extends DialogWrapper {
                 conditionsCheckboxList.updateItem(selectedCondition, newCondition, newCondition.asReadableString(FIELD_SEPARATOR));
                 newCondition.setEnabled(isEnabled);
             }
-        }).setRemoveAction(anActionButton -> {
-            ListUtil.removeSelectedItems(conditionsCheckboxList);
-        }).setButtonComparator("Add", "Edit", "Remove").createPanel();
+        }).setRemoveAction(anActionButton ->
+            ListUtil.removeSelectedItems(conditionsCheckboxList)
+        ).setButtonComparator("Add", "Edit", "Remove").createPanel();
     }
 
     /**
@@ -216,5 +266,57 @@ public class ModelDialog extends DialogWrapper {
             return new ValidationInfo("Please add a condition to your model!", toolbarPanel);
         }
         return super.doValidate();
+    }
+
+    private static class ComboBoxRenderer extends JLabel implements ListCellRenderer<Object> {
+
+        @SuppressWarnings("OverridableMethodCallInConstructor")
+        ComboBoxRenderer() {
+            setOpaque(true);
+            setHorizontalAlignment(LEFT);
+            setVerticalAlignment(CENTER);
+        }
+
+        @Override
+        public Component getListCellRendererComponent(JList list, Object value, int index, boolean isSelected, boolean cellHasFocus) {
+            String text = null;
+            Icon icon = null;
+            if (value instanceof BundledIcon) {
+                text = ((BundledIcon) value).getDescription();
+                icon = IconLoader.getIcon(((BundledIcon) value).getIconPath(), CustomIconLoader.class);
+            } else if (value instanceof String) {
+                text = (String) value;
+                icon = new ImageIcon();
+            }
+            setText(text);
+            setIcon(icon);
+
+            if (isSelected) {
+                setBackground(list.getSelectionBackground());
+                setForeground(list.getSelectionForeground());
+            } else {
+                setBackground(list.getBackground());
+                setForeground(list.getForeground());
+            }
+            return this;
+        }
+    }
+
+    private static class BundledIcon {
+        private final String iconPath;
+        private final String description;
+
+        public BundledIcon(String iconPath, String description) {
+            this.iconPath = iconPath;
+            this.description = description;
+        }
+
+        public String getIconPath() {
+            return iconPath;
+        }
+
+        public String getDescription() {
+            return description;
+        }
     }
 }